feat(win): reuse d2d
